+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add heart inside the now playing widget
 * Add heart icon to side nav
 * Fixes to loved actions not in sync
+* Security fixes
 
 ### 1.0.0: 2025-07-02